modelserver/requirements.txt
redis
Pillow
tensorflow-gpu==2.4.0
opencv-python
numpy
scipy
